% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mick-package.R
\docType{package}
\name{mick-package}
\alias{mick}
\alias{mick-package}
\title{mick: Mosaic Intelligent Clustering Kernel}
\description{
Tools for reducing high-resolution spatial and temporal energy-system input
data into tractable representative models.
}
\keyword{internal}
